apps/rag/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
langchain==0.3.25
langchain-openai==0.3.18
langchain-community==0.3.24
langchain-google-genai==2.1.5
psycopg2-binary==2.9.10
pgvector==0.4.1
pdfplumber==0.11.6
python-mammoth==1.9.0
beautifulsoup4==4.13.4
httpx==0.28.1
python-dotenv==1.1.0
python-multipart==0.0.20
